Add a <meta-data android:name="replacement" /> option
Allow developers to mark their MuzeiArtSource as having an available replacement MuzeiArtProvider. This will be used when Muzei upgrades to properly select the correct provider by default, is used to prevent legacy sources from appearing in the LegacyArtProvider, and makes it possible to simultaneously support older versions of Muzei as well as newer versions.

Also officially deprecates usage of MuzeiArtSource and RemoteMuzeiArtSource, pointing to MuzeiArtProvider as the replacement.